refactor: improve CSS & HTML structure
