--- client/src/assets/styles/boutonAccueil.css
@@ -1,21 +1,0 @@
-.logo {
-  left: 30%;
-  border-radius: 9999px;
-  height: 25px;
-  width: 25px;
-  border: none;
-  cursor: pointer;
-}
-
-.logo:hover {
-  box-shadow: 0 0 20px red;
-}
-
-@media screen and (min-width: 769px) {
-  .logo {
-    position: fixed;
-    left: 80px;
-    height: 30px;
-    width: 30px;
-  }
-}
