+++ client/src/components/LocationInput.css
@@ -17,12 +17,3 @@ input {
   border: 1px solid #ccc;
   border-radius: 4px;
 }
-
-button {
-  padding: 0.4rem 1rem;
-  background-color: #0066cc;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
