+++ client/src/components/boutonAccueil.tsx
@@ -1,5 +1,4 @@
 import { useNavigate } from "react-router";
-import logo from "../assets/icon/logo maison.png";
 
 function BoutonAccueil() {
   const navigate = useNavigate();
@@ -8,11 +7,17 @@ function BoutonAccueil() {
     navigate("/");
   }
   return (
-    <div>
-      <button onClick={handleclick} type="button">
-        <img className="logo" src={logo} alt="logo" />
-      </button>
-    </div>
+    <button onClick={handleclick} type="button">
+      <svg
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        width="2em"
+        height="2em"
+        viewBox="0 0 24 24"
+      >
+        <path fill="currentColor" d="M10 20v-6h4v6h5v-8h3L12 3L2 12h3v8z" />
+      </svg>
+    </button>
   );
 }
