+++ client/src/components/boutonConnexion.tsx
@@ -3,11 +3,7 @@ import { Link } from "react-router";
 function BoutonDesktop() {
   return (
     <nav>
-      <Link to={"/connexion"}>
-        <button className="connexion" type="button">
-          connexion
-        </button>
-      </Link>
+      <Link to={"/connexion"}>connexion</Link>
     </nav>
   );
 }
