+++ client/src/components/navigation.css
@@ -1,0 +1,38 @@
+.menu-nav {
+  position: fixed;
+  top: 30px;
+  left: 0;
+  right: 0;
+  background-color: rgba(0, 0, 0, 0.4);
+  width: 100%;
+  height: 60px;
+  box-shadow: -2px 0 10px rgba(0, 0, 0, 0.2);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px;
+  z-index: 999;
+  ul.liens {
+    position: fixed;
+    left: 0;
+    right: 200px;
+    top: 0;
+    bottom: 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 3rem;
+    list-style: none;
+    text-align: center;
+    gap: 60px;
+    background-color: white;
+    font-size: 21px;
+    @media screen and (min-width: 769px) {
+      position: absolute;
+      right: 0;
+      top: 60px;
+      bottom: auto;
+    }
+  }
+}
