+++ client/src/pages/MapComponent.css
@@ -2,7 +2,7 @@
 .carte-container {
   min-height: 100vh;
   padding: 100px 20px 80px; /* place pour navbar + footer */
-  background-image: url("../assets/images/background.webp"); /* mets le bon nom */
+  background-image: url("../assets/images/aae341cc-0264-4f21-a1c0-e449bb399e84.png"); /* mets le bon nom */
   background-size: cover;
   background-position: center;
   color: white;
@@ -17,6 +17,15 @@
   margin-bottom: 20px;
 }
 
+.carte-content {
+  display: grid;
+  grid-template-columns: 1fr; /* une seule colonne sur mobile */
+  gap: 1rem;
+  @media (min-width: 768px) {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
 .map-container {
   margin: 0 auto 40px;
   max-width: 800px;
@@ -27,10 +36,14 @@
 
 /* Composant List */
 .artwork-list {
+  order: 1;
   display: flex;
   flex-direction: column;
   gap: 20px;
   align-items: center;
+  @media (min-width: 768px) {
+    order: 0; /* pour que la liste soit à gauche sur les écrans plus larges */
+  }
 }
 
 .artwork-card {
