+++ client/src/pages/formulaire.css
@@ -1,24 +1,3 @@
-/* Variables CSS globales */
-:root {
-  --font-family: "retro-gaming", monospace;
-  --color-bg: #f0f0f0;
-  --color-text: #333;
-  --color-primary: rgb(0, 255, 187);
-  --color-primary-focus-shadow: rgba(0, 255, 187, 0.6);
-  --color-dark: #222;
-  --color-light: white;
-  --color-btn-hover: silver;
-
-  --border-radius: 10px;
-  --transition: 0.3s ease;
-}
-
-/* Styles globaux */
-
-body {
-  font-family: var(--font-family);
-}
-
 .profil {
   background-color: var(--color-light);
   padding: 1.5rem 1.5rem;
